+++ common-components/CHANGELOG.md
@@ -4,6 +4,12 @@ All changes to this project will be documented in this file.
 
 ## Template [MajorVersion.MediterraneanVersion.MinorVersion] - DD-MM-YYYY
 
+## [0.0.64] - 17.07.2026
+
+- Fixed Long Url's column ui
+- Updated sanitize-html package to latest version
+- Added yes/no filter to Comment and Feedback columns in the chat history table
+
 ## [0.0.63] - 06.07.2026
 
 - Modified toast position and appearance